# [BEAM-8832] Allow GCS staging upload chunk size to be increased >1M when setting GcsUploadBufferSizeBytes pipeline option, to improve performance