@lukecwik 2019-11-27
The limit of 1mb was being used to not blow out memory in the case when the user has 100s of jars.

I don't believe the package stager limits the number of concurrent writes so in this case the users code would start to use 16x more memory.