@alanmyrvold 2019-11-27
@lukecwik Updated the pull request to not change the default, but allow the user specified GcsUploadBufferSizeBytes pipeline option to exceed the 1M value, which keeps the same behavior for users not specifying the option.